data delete with sweetalert updated anmd data added feature done

+++ src/context/notes/noteContext.js
@@ -1,5 +1,9 @@
 import React, {createContext, useState} from 'react';
 import http from "../../services/httpRequest";
+import {showSuccessToast} from "../../services/toastServices";
+import swal from "sweetalert2";
+import axios from "axios";
+import API_BASE_URL from "../../config";
 
 export const NoteContext = createContext();
 
@@ -11,18 +15,18 @@ export const NoteProvider = ({children}) => {
         setNotes(response);
     }
 
-    const addNote = (noteObj) => {
+    const addNote = async (noteObj) => {
         const {title, description, tag} = noteObj;
-        const note = {
-            "_id": "64366e6170cb6e235aaeb3bd1",
-            "user": "643622fdca7bf7a6b0c8662d",
-            "title": title,
-            "description": description,
-            "tag": tag,
-            "date": "2023-04-12T08:40:01.772Z",
-            "__v": 0
+        const note = {title, description, tag}
+        try {
+            const response = await http.post("notes", note);
+            console.log(response);
+            const allNotes = [response, ...notes];
+            setNotes(allNotes);
+            showSuccessToast("Note Added Successfully!");
+        } catch (e) {
+            console.error(e.message)
         }
-        setNotes(notes.concat(note));
     }
 
     const editNote = (noteId, noteObj) => {
@@ -40,10 +44,27 @@ export const NoteProvider = ({children}) => {
 
     const deleteNote = async (noteId) => {
         const allNotes = [...notes];
-        const filteredNotes = notes.filter(note => note._id !== noteId);
-        setNotes(filteredNotes);
         try {
-            await http.delete("notes", noteId);
+            swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    const filteredNotes = notes.filter(note => note._id !== noteId);
+                    setNotes(filteredNotes);
+                    await http.delete("notes", noteId);
+                    swal.fire(
+                        'Deleted!',
+                        'Your file has been deleted!',
+                        'success'
+                    )
+                }
+            });
         } catch (e) {
             console.error(e.message)
             setNotes(allNotes);
